[FT0] Rename DigitizationParameters -> FT0DigParam
