+++ Detectors/FIT/FT0/reconstruction/src/CTFCoder.cxx
@@ -14,7 +14,7 @@
 /// \brief class for entropy encoding/decoding of FT0 digits data
 
 #include "FT0Reconstruction/CTFCoder.h"
-#include "FT0Simulation/DigitizationParameters.h"
+#include "FT0Simulation/FT0DigParam.h"
 #include "CommonUtils/StringUtils.h"
 #include <TTree.h>
 
@@ -51,7 +51,7 @@ void CTFCoder::compress(CompressedDigits& cd, const gsl::span<const Digit>& digi
   cd.header.nTriggers = digitVec.size();
   cd.header.firstOrbit = dig0.getOrbit();
   cd.header.firstBC = dig0.getBC();
-  cd.header.triggerGate = DigitizationParameters::Instance().mTime_trg_gate;
+  cd.header.triggerGate = FT0DigParam::Instance().mTime_trg_gate;
 
   cd.trigger.resize(cd.header.nTriggers);
   cd.bcInc.resize(cd.header.nTriggers);
