+++ Detectors/FIT/FT0/simulation/include/FT0Simulation/Digitizer.h
@@ -22,7 +22,7 @@
 #include "SimulationDataFormat/MCTruthContainer.h"
 #include "SimulationDataFormat/MCCompLabel.h"
 #include "FT0Simulation/DigitizationConstants.h"
-#include "FT0Simulation/DigitizationParameters.h"
+#include "FT0Simulation/FT0DigParam.h"
 #include <TH1F.h>
 #include <array>
 #include <bitset>
@@ -103,7 +103,7 @@ class Digitizer
     if (x <= 0.0f) {
       return 0.0f;
     }
-    float const y = x / DigitizationParameters::Instance().mBunchWidth * DP::SIGNAL_TABLE_SIZE;
+    float const y = x / FT0DigParam::Instance().mBunchWidth * DP::SIGNAL_TABLE_SIZE;
     int const index = std::floor(y);
     if (index + 1 >= DP::SIGNAL_TABLE_SIZE) {
       return mSignalTable.back();
@@ -116,7 +116,7 @@ class Digitizer
   inline VcType signalFormVc(VcType x) const
   { // table lookup for the signal shape (SIMD version)
     // implemented as template function, so that we don't need to include <Vc/Vc> here
-    auto const y = x / DigitizationParameters::Instance().mBunchWidth * DP::SIGNAL_TABLE_SIZE;
+    auto const y = x / FT0DigParam::Instance().mBunchWidth * DP::SIGNAL_TABLE_SIZE;
     typename VcType::IndexType const index = floor(y);
     auto const rem = y - index;
     VcType val(0);
