+++ Detectors/FIT/FT0/simulation/src/FT0SimulationLinkDef.h
@@ -17,8 +17,9 @@
 
 #pragma link C++ class o2::ft0::Detector + ;
 #pragma link C++ class o2::ft0::Digitizer + ;
-#pragma link C++ class o2::ft0::DigitizationParameters + ;
+#pragma link C++ class o2::ft0::FT0DigParam + ;
 #pragma link C++ class o2::base::DetImpl < o2::ft0::Detector> + ;
 #pragma link C++ class o2::ft0::Digits2Raw + ;
 
 #endif
+
